+++ edgelet/docker-rs/src/apis/old/image_api.rs
